extract MockFileUtils from Mocks class to prevent runtime issue resolving grails.test.MockUtils
